tts_dasarPemrogramanJaringan/be/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
marshmallow==3.20.1
python-dotenv==1.0.0
flask-cors==4.0.0
cryptography==41.0.4